+++ modules/settings/theme-settings.js
@@ -3,17 +3,7 @@ import { SettingsMenuDorakoUI } from "./menu.js";
 export class ThemeSettings extends SettingsMenuDorakoUI {
   static namespace = "theme";
 
-  static SETTINGS = [
-    "dark-theme-degree",
-    "npc-sheet-theme",
-    "pc-sheet-theme",
-    "loot-sheet-theme",
-    "familiar-sheet-theme",
-    "chat-theme",
-    "header-style",
-    "frosted-glass",
-    "glass-bg",
-  ];
+  static SETTINGS = ["application-theme", "chat-theme", "header-style", "frosted-glass", "glass-bg"];
 
   static get settings() {
     return {
@@ -84,78 +74,17 @@ export class ThemeSettings extends SettingsMenuDorakoUI {
           root.setProperty("--glass-bg", game.settings.get("pf2e-dorako-ui", "theme.glass-bg"));
         },
       },
-      "dark-theme-degree": {
-        name: "pf2e-dorako-ui.settings.theme.dark-theme-degree.name",
-        hint: "pf2e-dorako-ui.settings.theme.dark-theme-degree.hint",
-        scope: "client",
-        config: true,
-        default: "none",
-        type: String,
-        choices: {
-          none: "pf2e-dorako-ui.settings.theme.dark-theme-degree.choice.none",
-          supported: "pf2e-dorako-ui.settings.theme.dark-theme-degree.choice.supported",
-          extended: "pf2e-dorako-ui.settings.theme.dark-theme-degree.choice.extended",
-          maximum: "pf2e-dorako-ui.settings.theme.dark-theme-degree.choice.maximum",
-        },
-        requiresReload: false, // re-render all windows
-        onChange: () => {
-          const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
-          for (const app of apps) {
-            app.render();
-          }
-        },
-      },
-      "pc-sheet-theme": {
-        name: "pf2e-dorako-ui.settings.theme.pc-sheet-theme.name",
-        hint: "pf2e-dorako-ui.settings.theme.pc-sheet-theme.hint",
-        scope: "client",
-        config: true,
-        default: "default",
-        type: String,
-        choices: {
-          default: "pf2e-dorako-ui.text.default",
-          // "light-theme": "pf2e-dorako-ui.text.light",
-          "dark-theme": "pf2e-dorako-ui.text.dark",
-        },
-        requiresReload: false, // re-render all windows
-        onChange: () => {
-          const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
-          for (const app of apps) {
-            app.render();
-          }
-        },
-      },
-      "npc-sheet-theme": {
-        name: "pf2e-dorako-ui.settings.theme.npc-sheet-theme.name",
-        hint: "pf2e-dorako-ui.settings.theme.npc-sheet-theme.hint",
+      "application-theme": {
+        name: "pf2e-dorako-ui.settings.theme.application-theme.name",
+        hint: "pf2e-dorako-ui.settings.theme.application-theme.hint",
         scope: "client",
         config: true,
         default: "light-theme",
         type: String,
         choices: {
-          default: "pf2e-dorako-ui.text.default",
-          "light-theme": "pf2e-dorako-ui.text.light",
-          "dark-theme": "pf2e-dorako-ui.text.dark",
-        },
-        requiresReload: false, // re-render all windows
-        onChange: () => {
-          const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
-          for (const app of apps) {
-            app.render();
-          }
-        },
-      },
-      "familiar-sheet-theme": {
-        name: "pf2e-dorako-ui.settings.theme.familiar-sheet-theme.name",
-        hint: "pf2e-dorako-ui.settings.theme.familiar-sheet-theme.hint",
-        scope: "client",
-        config: true,
-        default: "default",
-        type: String,
-        choices: {
-          default: "pf2e-dorako-ui.text.default",
-          dark: "pf2e-dorako-ui.text.dark",
-          darkRedHeader: "pf2e-dorako-ui.settings.theme.familiar-sheet-theme.choice.dark-red-header",
+          "no-theme": "pf2e-dorako-ui.settings.theme.application-theme.choice.no-theme",
+          "light-theme": "pf2e-dorako-ui.settings.theme.application-theme.choice.light-theme",
+          "dark-theme": "pf2e-dorako-ui.settings.theme.application-theme.choice.dark-theme",
         },
         requiresReload: false, // re-render all windows
         onChange: () => {
@@ -166,26 +95,89 @@ export class ThemeSettings extends SettingsMenuDorakoUI {
         },
       },
 
-      "loot-sheet-theme": {
-        name: "pf2e-dorako-ui.settings.theme.loot-sheet-theme.name",
-        hint: "pf2e-dorako-ui.settings.theme.loot-sheet-theme.hint",
-        scope: "client",
-        config: true,
-        default: "light-theme",
-        type: String,
-        choices: {
-          default: "pf2e-dorako-ui.text.default",
-          "light-theme": "pf2e-dorako-ui.text.light",
-          // "dark-theme": "pf2e-dorako-ui.text.dark",
-        },
-        requiresReload: false, // re-render all windows
-        onChange: () => {
-          const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
-          for (const app of apps) {
-            app.render();
-          }
-        },
-      },
+      // "pc-sheet-theme": {
+      //   name: "pf2e-dorako-ui.settings.theme.pc-sheet-theme.name",
+      //   hint: "pf2e-dorako-ui.settings.theme.pc-sheet-theme.hint",
+      //   scope: "client",
+      //   config: true,
+      //   default: "default",
+      //   type: String,
+      //   choices: {
+      //     default: "pf2e-dorako-ui.text.default",
+      //     // "light-theme": "pf2e-dorako-ui.text.light",
+      //     "dark-theme": "pf2e-dorako-ui.text.dark",
+      //   },
+      //   requiresReload: false, // re-render all windows
+      //   onChange: () => {
+      //     const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
+      //     for (const app of apps) {
+      //       app.render();
+      //     }
+      //   },
+      // },
+
+      // "npc-sheet-theme": {
+      //   name: "pf2e-dorako-ui.settings.theme.npc-sheet-theme.name",
+      //   hint: "pf2e-dorako-ui.settings.theme.npc-sheet-theme.hint",
+      //   scope: "client",
+      //   config: true,
+      //   default: "light-theme",
+      //   type: String,
+      //   choices: {
+      //     default: "pf2e-dorako-ui.text.default",
+      //     "light-theme": "pf2e-dorako-ui.text.light",
+      //     "dark-theme": "pf2e-dorako-ui.text.dark",
+      //   },
+      //   requiresReload: false, // re-render all windows
+      //   onChange: () => {
+      //     const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
+      //     for (const app of apps) {
+      //       app.render();
+      //     }
+      //   },
+      // },
+
+      // "familiar-sheet-theme": {
+      //   name: "pf2e-dorako-ui.settings.theme.familiar-sheet-theme.name",
+      //   hint: "pf2e-dorako-ui.settings.theme.familiar-sheet-theme.hint",
+      //   scope: "client",
+      //   config: true,
+      //   default: "default",
+      //   type: String,
+      //   choices: {
+      //     default: "pf2e-dorako-ui.text.default",
+      //     dark: "pf2e-dorako-ui.text.dark",
+      //     darkRedHeader: "pf2e-dorako-ui.settings.theme.familiar-sheet-theme.choice.dark-red-header",
+      //   },
+      //   requiresReload: false, // re-render all windows
+      //   onChange: () => {
+      //     const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
+      //     for (const app of apps) {
+      //       app.render();
+      //     }
+      //   },
+      // },
+
+      // "loot-sheet-theme": {
+      //   name: "pf2e-dorako-ui.settings.theme.loot-sheet-theme.name",
+      //   hint: "pf2e-dorako-ui.settings.theme.loot-sheet-theme.hint",
+      //   scope: "client",
+      //   config: true,
+      //   default: "light-theme",
+      //   type: String,
+      //   choices: {
+      //     default: "pf2e-dorako-ui.text.default",
+      //     "light-theme": "pf2e-dorako-ui.text.light",
+      //     // "dark-theme": "pf2e-dorako-ui.text.dark",
+      //   },
+      //   requiresReload: false, // re-render all windows
+      //   onChange: () => {
+      //     const apps = Object.values(ui.windows).filter((w) => w instanceof Application);
+      //     for (const app of apps) {
+      //       app.render();
+      //     }
+      //   },
+      // },
     };
   }
 }